+++ apiphany-core/src/test/java/org/apiphany/security/ssl/SSLContextsTest.java
@@ -17,7 +17,7 @@ import java.security.UnrecoverableKeyException;
 
 import javax.net.ssl.SSLContext;
 
-import org.apiphany.io.FileSource;
+import org.apiphany.io.ResourceLocation;
 import org.apiphany.json.JsonBuilder;
 import org.apiphany.lang.Strings;
 import org.apiphany.test.Assertions;
@@ -71,7 +71,7 @@ class SSLContextsTest {
 	@ParameterizedTest
 	@MethodSource("provideEmptyPasswords")
 	void shouldLoadKeyStoreWithEmptyPasswordButFailToLoadKey(final char[] password) throws KeyStoreException {
-		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, password, FileSource.CLASS_PATH);
+		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, password, ResourceLocation.CLASS_PATH);
 
 		assertThat(keyStore.size(), equalTo(1));
 
@@ -94,7 +94,8 @@ class SSLContextsTest {
 	void shouldFailToLoadKeyStoreWithWrongPassword() {
 		char[] wrongPassword = "wrongpassword".toCharArray();
 		SecurityException exception =
-				assertThrows(SecurityException.class, () -> SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, wrongPassword, FileSource.CLASS_PATH));
+				assertThrows(SecurityException.class,
+						() -> SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, wrongPassword, ResourceLocation.CLASS_PATH));
 
 		assertNotNull(exception);
 		assertThat(exception.getMessage(), equalTo("Error loading key store: " + KEYSTORE_PATH));
@@ -107,7 +108,7 @@ class SSLContextsTest {
 	@Test
 	void shouldLoadKeyStoreSuccessfully() throws KeyStoreException {
 		char[] correctPassword = "keystorepassword123".toCharArray();
-		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, correctPassword, FileSource.CLASS_PATH);
+		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, correctPassword, ResourceLocation.CLASS_PATH);
 
 		assertThat(keyStore.size(), equalTo(1));
 	}
@@ -120,7 +121,7 @@ class SSLContextsTest {
 
 		char[] correctPassword = "keystorepassword123".toCharArray();
 
-		KeyStore keyStore = SSLContexts.keyStore(absolutePath.toString(), KEYSTORE_TYPE, correctPassword, FileSource.FILE_SYSTEM);
+		KeyStore keyStore = SSLContexts.keyStore(absolutePath.toString(), KEYSTORE_TYPE, correctPassword, ResourceLocation.FILE_SYSTEM);
 
 		assertThat(keyStore.size(), equalTo(1));
 	}
@@ -130,7 +131,7 @@ class SSLContextsTest {
 		String invalidPath = "invalid/path/to/keystore.jks";
 
 		SecurityException exception =
-				assertThrows(SecurityException.class, () -> SSLContexts.keyStore(invalidPath, KEYSTORE_TYPE, null, FileSource.CLASS_PATH));
+				assertThrows(SecurityException.class, () -> SSLContexts.keyStore(invalidPath, KEYSTORE_TYPE, null, ResourceLocation.CLASS_PATH));
 
 		assertThat(exception.getMessage(), equalTo("Error loading key store: " + invalidPath));
 	}
@@ -140,42 +141,42 @@ class SSLContextsTest {
 		String invalidPath = "/invalid/absolute/path/to/keystore.jks";
 
 		SecurityException exception =
-				assertThrows(SecurityException.class, () -> SSLContexts.keyStore(invalidPath, KEYSTORE_TYPE, null, FileSource.FILE_SYSTEM));
+				assertThrows(SecurityException.class, () -> SSLContexts.keyStore(invalidPath, KEYSTORE_TYPE, null, ResourceLocation.FILE_SYSTEM));
 
 		assertThat(exception.getMessage(), equalTo("Error loading key store: " + invalidPath));
 	}
 
 	@Test
 	void shouldLoadKeyWithNullPasswordIfAllowed() throws KeyStoreException {
-		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, null, FileSource.CLASS_PATH);
+		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, null, ResourceLocation.CLASS_PATH);
 
 		assertThat(keyStore.size(), equalTo(1));
 	}
 
 	@Test
 	void shouldLoadKeyWithEmptyPasswordIfAllowed() throws KeyStoreException {
-		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, new char[] { }, FileSource.CLASS_PATH);
+		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, KEYSTORE_TYPE, new char[] { }, ResourceLocation.CLASS_PATH);
 
 		assertThat(keyStore.size(), equalTo(1));
 	}
 
 	@Test
 	void shouldReturnNullIfKeyStoreTypeIsEmpty() {
-		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, "", null, FileSource.CLASS_PATH);
+		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, "", null, ResourceLocation.CLASS_PATH);
 
 		assertNull(keyStore);
 	}
 
 	@Test
 	void shouldReturnNullIfKeyStoreTypeIsNull() {
-		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, null, null, FileSource.CLASS_PATH);
+		KeyStore keyStore = SSLContexts.keyStore(KEYSTORE_PATH, null, null, ResourceLocation.CLASS_PATH);
 
 		assertNull(keyStore);
 	}
 
 	@Test
 	void shouldReturnNullIfKeyStoreLocationIsEmpty() {
-		KeyStore keyStore = SSLContexts.keyStore("", KEYSTORE_TYPE, null, FileSource.CLASS_PATH);
+		KeyStore keyStore = SSLContexts.keyStore("", KEYSTORE_TYPE, null, ResourceLocation.CLASS_PATH);
 
 		assertNull(keyStore);
 	}